refactor(421): extract ADT memory-layout utils to vera/codegen/memory.py
Step 1 of the codegen/api.py decomposition (#421): the pure ADT memory-layout utilities (ConstructorLayout, _validate_wrap_handle, _wasm_type_size/_wasm_type_align/_align_up) are lifted out of the 4,358-line api.py into a focused vera/codegen/memory.py with no wasmtime/runtime dependency. Consumers (codegen/__init__, core, registration, closures, and the _validate_wrap_handle unit tests) import them from memory.py directly; api.py keeps only _validate_wrap_handle, which _wrap_handle uses internally inside execute().

Public surface unchanged: 'from vera.codegen import ConstructorLayout' and 'from vera.codegen.api import compile/execute/CompileResult/ExecuteResult' all resolve. api.py 4,358->4,247 lines. mypy + ruff clean; the #734 characterization harness (22 tests) and the moved-fn tests (6) stay green.
