+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - **Mutation-testing infrastructure and soundness-core baseline** ([#387](https://github.com/aallan/vera/issues/387)).  Adds `mutmut` as a `[mutation]` extra (chosen over `cosmic-ray` after a Phase-0 spike — its coverage-guided test selection is ~90× faster on this Z3/WASM suite, and its copy-based model never mutates the working tree in place), a `[tool.mutmut]` config wiring the **in-process unit-test oracle** (the subprocess suites `test_conformance` / `test_cli` / `test_browser` import the un-mutated package, so they cannot kill mutants), and a runbook (`MUTATION.md`) covering the tool decision, the oracle caveat, resume-after-hard-kill, the Z3-flakiness guardrail, and the equivalent-mutant `# pragma: no mutate` convention.  The first sweep measures the **soundness core** (`verifier.py`, `smt.py`, `checker/`, `obligations/` — 10,620 mutants): **80.8% caught**, 2,038 survivors, emitted by `scripts/mutation_report.py` as a diff-able per-module `mutation-summary.csv` and a README mutation-score badge (`mutation.json`), with the full survivor inventory and per-module chart attached to the tracking issue.  The whole-`vera/` sweep is deferred behind [#421](https://github.com/aallan/vera/issues/421): `codegen/api.py`'s ~3,500-line `execute()` inflates a 774 MB mutant file mutmut cannot index.  Tooling and tests only — no behaviour change; the issue stays open for soundness-core triage and per-module follow-ups.
 - **Soundness-core test hardening (Phase 3 of [#387](https://github.com/aallan/vera/issues/387))**.  Strengthens type-checker tests in the call-, control-, and expression-checking core (`vera/checker/calls.py`, `vera/checker/control.py`, `vera/checker/core.py`, `vera/checker/expressions.py`) that stayed *green* when behaviour was deliberately broken under mutation: each hardened or added test now asserts a **specific observable** — the stable diagnostic `error_code` (call/constructor `E203`–`E204`, `E210`–`E215`, `E220`, `E230`–`E233`, `E240`–`E241`; match-arm `E302`; exhaustiveness `E310`–`E313`; pattern `E320`–`E322`; handler `E330`–`E335` plus body-type `E121`; data-invariant `E120`; expression `E146`, `E160`–`E161`, `E173`–`E175`), the diagnostic `severity`, or a type-propagation result — rather than free-text message prose.  Test-only; no behaviour change.  The tracking issue stays open for the whole-corpus sweep gated on [#421](https://github.com/aallan/vera/issues/421).
 
+### Changed
+
+- **`codegen/api.py` decomposition — step 1: ADT memory layout ([#421](https://github.com/aallan/vera/issues/421))**.  Lifted the pure ADT memory-layout utilities (`ConstructorLayout`, `_validate_wrap_handle`, `_wasm_type_size` / `_wasm_type_align` / `_align_up`) out of the 4,358-line `vera/codegen/api.py` into a focused `vera/codegen/memory.py`, with the codegen consumers importing them from there directly.  The public surface is unchanged — `from vera.codegen import ConstructorLayout` and `from vera.codegen.api import compile` / `execute` / `CompileResult` / `ExecuteResult` all still resolve — and the #734 characterization harness stays green.  First of a per-concern series that lifts the runtime out of the public-API module so the deferred whole-`vera/` mutation sweep ([#387](https://github.com/aallan/vera/issues/387)) becomes tractable.  Internal refactor; no behaviour change.
+
 ## [0.0.177] - 2026-06-21
 
 ### Added
